Move create wallets to module
